Restart the nginx server

+++ 2-do_deploy_web_static.py
@@ -12,8 +12,10 @@ def do_deploy(archive_path):
 
     if not isfile(archive_path):
         return False
-    put(archive_path, '/tmp/', use_sudo=True, mirror_local=True)
-    try:
+    print(type(archive_path))
+    if not put(local_path=archive_path, remote_path='/tmp/').succeeded:
+        return False
+    """try:
         sudo('tar -xvzf /tmp/web_static_*.tgz -C /data/web_static/releases/')
         sudo('rm /tmp/web_static_*.tgz /data/web_static/current')
         run(
@@ -21,7 +23,7 @@ def do_deploy(archive_path):
                     /data/web_static/current')
         return True
     except Exception:
-        return False
+        return False"""
 
 
-env.hosts = ['54.87.205.95', '54.237.226.36']
+env.hosts = ['54.237.226.36']
